fuzz: initial fuzzing infrastructure, docs and and Base64::encode() e… (#2476)
This is the groundwork for #508. See test/fuzz/README.md for the spiel
on what we're integrating with, the fuzz test model and how to create
new fuzz tests.

There were a bunch of changes needed to the build infrastructure to get
it to work in the oss-fuzz Docker build environment. In particular, we
needed to allow clang + libc++ static linking on Linux and to supply
extra flags that some of the external deps were not handling correctly.

Signed-off-by: Harvey Tuch <htuch@google.com>