upstream: encapsulate hosts per locality in class, type cleanups. (#2516)
Rather than pass a std::vector<std::vector<HostSharedPtr>> around to
represent the hosts grouped by locality, switch to a first class object,
HostsPerLocality.

This has the advantage that we can independently flag whether or not the
local locality is present, which is needed to correct a bug in load
reporting where we fail to report for clusters where there are no local
hosts. The bug fix will be in a following PR.

The HostsPerLocality internal representation and API is fairly similar
to what we have today, a list with the first locality local (if it
exists). This has the advantage of minimizing churn and being largely
compatible with the existing zone aware LB. The interface is immutable
however, which reduces the amount of reliance code has on the internal
representation.

Risk Level: Low
Testing: New UTs for HostsPerLocalityImpl.

Signed-off-by: Harvey Tuch <htuch@google.com>