server: factor out MainCommon as a class, with a run method (#2568)
This is step 2 in the plan to improve startup performance. It makes it possible instantiate the full server stack without running the event loop, so the instantiation can be timed. This also simplifies the code required in main(), though a legacy path is retained allowing existing alternate main()s to continue to work.

Risk Level: Medium: this has been extensively tested but deserves scrutiny as it is part of the startup path.
Testing: //test/... including sanitizers.
Added tests for the new MainCommon instantiation, as well as the legacy main_common() where some of the work was delegated to main().

This is another step toward addressing #2373

Signed-off-by: Joshua Marantz <jmarantz@google.com>