server: make the --disable-hot-restart flag functional (#2679)
The --disable-hot-restart flag was added to the binary in #2576 by @tonya11en but not hooked up (at my request). Now that the main_common refactor is in, it's easier to hook it up. This also offers the opportunity for some minor code simplifications.

Risk Level: Medium -- this affects the startup path.

Testing:
//test/...
added a new subtest to hotrestart_test.sh to cover --disable-hot-restart. I'd appreciate suggestions though on how to cover the functionality more thoroughly...how should I detect from a test that hot restart is disabled?

Release Notes: TBD: I'm not sure if release notes were added already in #2576 -- if not they can be added during this PR review.

*Fixes #2029 -- actually #2576 claims to fix that but I don't think it really does, since in that PR the switch is a no-op.

Signed-off-by: Joshua Marantz <jmarantz@google.com>