eds: allow removing all priorities (#3354)
While writing some additional tests for #3327 I ran into an issue with clearing out all the priorities. This fixes this bug by only ever adding an element to priority_state when we see an actual endpoint. This allows clearing out all priorities for a cluster, which is something that we need to be able to do on our end in order to fully disable a cluster.

Without this fix, attempts to clear out priority 0 would have no effect, since the secondary loop that cleans out all priorities that aren't in the update was starting at priority 1.

The test case added here also serves to prove that the bug reported in #3327 actually works as intended: a locality can be removed from a priority simply by not referencing it in the EDS config update.

Risk Level:
Low: Small bug fix

Signed-off-by: Snow Pettersen <snowp@squareup.com>