http: better handling of HEAD requests (#3775)
Right now, when Envoy receives a response to a HEAD request it applies standard handling of header-only requests and appends a "content-length: 0" header. By the standard* headers for HEAD responses should match the headers for GET responses. This PR brings that about for all cases but one: because Envoy can not infer the actual content length from upstream, if there is no content length or transfer encoding header, but there is a 0 byte body, Envoy will guess there is a non-zero body and still append a transfer-encoding: chunked header. We could fix that by always sending transfer-encoding chunked even if we know it's a header only response, but I think we can leave that until-if someone complains.

*https://www.w3.org/Protocols/rfc2616/rfc2616-sec9.html#sec9.4

Risk Level: High (messing with existing HTTP header processing)
Testing: new integration tests and unit tests for HEAD requests
Docs Changes: n/a
Release Notes: inline

Signed-off-by: Alyssa Wilk <alyssar@chromium.org>