network: fixing a bug in detecting closed connections (#3813)
As documented in #3785, we had a disconnect test flaking out because the ECONNRESET was being overwritten by EAGAIN in the logging code, so the ECONNRESET was not properly handled and the test hung. This represents an actual Envoy bug where an inopportune log flush can mask a connection close. Presumably the same issue exists on the write path.

Risk Level: Low
Testing: IpVersions/Http2IntegrationTest.RouterDownstreamDisconnectBeforeRequestComplete/IPv6 less flaky
Docs Changes: n/a
Release Notes: n/a
Fixes #3785

Signed-off-by: Alyssa Wilk <alyssar@chromium.org>