syscall: refactor address APIs for deeper errno latching (#3897)
* Refactor address APIs for deeper errno latching

The errno set by a syscall can be overwritten by code (e.g. logging) as
it propagates up through the call stack. This commit refactors the
bind and connect methods in the address API to allow for returning the
errno from deeper down the call stack i.e. as soon as a syscall is
performed.

Signed-off-by: Venil Noronha <veniln@vmware.com>