test: Fix inverted exact match logic in IntegrationTcpClient::waitForData() (#4134)
The initial check in IntegrationTcpClient::waitForData() did exact
match check when 'exact_match' parameter was 'false', and an inexact
match when 'exact_match' parameter was 'true'. If the initial test
fails, the forthcoming test in WaitForPayloadReader::onData() was
using the logically correct interpretation of the 'exact_match'
parameter.

This resulted in wait never completing for inexact matches
that should have matched data that was already received. The correct
test in WaitForPayloadReader::onData() would always correctly match
data that is received after IntegrationTcpClient::waitForData() is
called.

Fix the incorrect logic by reuing the logical pattern of
WaitForPayloadReader::onData() also in
IntegrationTcpClient::waitForData(), except for not skipping the exact
match on an empty search string.

Testing: Added a iine in TCP proxy integration test that fails/hangs
without the fix.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>