alts: add an unused implementation NoOpTransportSocketCallbacks (#4181)
Add an implementation of TransportSocketCallbacks which does nothing when setReadBufferReady() or raiseEvent() is call. This implementation is supposed to be used in TransportSocket implementation which wraps another socket object. In this case, the underlying socket's callbacks should be suppressed.

Also add callbacks() in TransportSocket interface which is used to get the callbacks passed to setTransportSocketCallbacks().

Risk Level: Low (alts utility, not enabled in main)
Testing: new unit tests
Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Dan Zhang <danzh@google.com>