Shared FilterState: List support (#4588)
This is an attempt to solve #4528

I tried adding a setList and getList or other variants to the filter state interface as @curiouserrandy suggested in the issue above. Two problems came up:

1. hasData method becomes complicated. It effectively becomes "check if all elements of the list have same type"
2. getListData also becomes kludgey. Obtain the vector, dynamic cast each element to check if typematches and then return a vector of references. Not sure if its worth the perf overhead to do this over and over again.

Instead, I opted here to return a non-const reference to the underlying object (which is not const either). Short of me doing my own const cast..

Signed-off-by: Shriram Rajagopalan <shriramr@vmware.com>