config: removing internal code structures allowing v1 config (#5552)
Removing v2ConfigOnly(). I think this is the removal for v1 support in Envoy.

Previously this was used to allow v1 config via command line switch. Currently it's only used in upstream Envoy to allow v1 config for the last lingering v1 tests.
intentionally not removing BootstrapJson::translateBootstrap until end of quarter as it's still used in Harvey's awesome v1_to_bootstrap util which we are not sunsetting until March or so.

Risk Level: Medium - no way to tell what downstream Envoys might still make use of this.
Testing: n/a
Docs Changes: n/a
Release Notes: not added.

Signed-off-by: Alyssa Wilk <alyssar@chromium.org>