upstream: move HostSet updates to PrioritySet (#5818)
This is a refactoring step to allow additional logic to be added when
determining what hosts are pass to the membership update callback.
Notably, this will allow EDS to process multiple priorities at once
without triggering intermediate membership update callbacks and allow
named/duplicated endpoints (#4280) to be used safely.

`HostSet` updating is moved to `PrioritySet`, which introduces a shared code
path that is used for all `HostSet` updates. `HostSet`s are now only exposed
as const refs, ensuring that they are not updated without going through
the `PrioritySet` interface. The key is moving the trigger for membership
update out of the `HostSet` so that multiple calls can be made to
`HostSetImpl::updateHost` without triggering an update.

Risk Level: Low, refactoring only
Testing: Current test suite

Signed-off-by: Snow Pettersen <snowp@squareup.com>