eds: use batched host set updates (#6082)
This changes EDS to use the newly introduced batched host update. This
ensures that all host set updates are applied before triggering
membership callbacks. The benefit of this is that connection pools
will no longer be drained if a host is moved between priorities.

Risk Level: Medium
Testing: Integration test

Signed-off-by: Snow Pettersen <snowp@squareup.com>