http timeout integration test: wait for 15s for upstream reset (#6646)
This test waits for the upstream to see a reset which confirms that the
router filter did the right thing when the global timeout is hit.
However since this involves the network, we would occasionally see the
reset after the wait call. Since we were waiting for 0ms we'd get
flakes. 15s is hopefully high enough that the test will succeed
reliably.

Signed-off-by: Michael Puncel <mpuncel@squareup.com>