Disable ranlib for cmake build dependencies (#6991)
`ranlib` is a tool to add a table of contents file to static archives.
On any modern system this is done automatically by creating the archive
with `ar` (unless you explicitly disable this behavior for performance
reasons). For backwards compatibility (I assume) cmake still calls
`ranlib` on archives to ensure this table of contents is created.
Unfortunately bazel knows nothing about ranlib, and never calls it for
any actions. Because of this rules_foreign_cc can not set the correct
path to `ranlib` based on the current `cc_toolchain`. Because of that
when you try to build envoy for the Android NDK, cmake auto-discovers a
version of ranlib, which on macOS happens to be `/usr/bin/ranlib`, which
then produces invalid ELF binaries.

Setting `CMAKE_RANLIB` to an empty string stops cmake from using it
ever, which avoids this issue.

Tracking upstream issue: https://github.com/bazelbuild/rules_foreign_cc/issues/252

Signed-off-by: Keith Smiley <keithbsmiley@gmail.com>