Add detekt for basic kotlin linting (#96)
https://github.com/lyft/envoy-mobile/issues/27

Adding a simple script and using openjdk:8-jdk Docker image. Open to moving some of the script dependencies to a Dockerfile and can create an issue for that if we are interested

Using https://github.com/arturbosch/detekt with the `--build-upon-default-config` flag which uses the [default config settings](https://github.com/arturbosch/detekt/blob/master/detekt-cli/src/main/resources/default-detekt-config.yml) and some customization on top (found in `.kotlinlint.yml`). This does not support `detekt-formatter` and that can be something we look into if we want to as well.

Signed-off-by: Alan Chiu <achiu@lyft.com>

For an explanation of how to fill out the fields, please see the relevant section
in [PULL_REQUESTS.md](https://github.com/envoyproxy/envoy/blob/master/PULL_REQUESTS.md)

Description: Add detekt for kotlin linting
Risk Level: low
Testing: locally and CI
Docs Changes: Added information about what we use to lint and the file we use
Release Notes:
[Optional Fixes #Issue]
[Optional Deprecated:]

Signed-off-by: JP Simard <jp@jpsim.com>