android: create network condition monitor (#468)
Description: Adds a platform network monitor to Android to leverage the set_preferred_network interface. This implementation only works on Lollipop and above, and does rely on some methods deprecated in the most recent versions of Android - in order to be compatible with multiple versions and not unnecessarily complicated.
Risk Level: low
Testing: local, CI, and physical device

Signed-off-by: Mike Schore <mike.schore@gmail.com>
Signed-off-by: JP Simard <jp@jpsim.com>