server: Introduce ServerFactoryContext (#8485)
Introduce ServerFactoryContext which is an implementation of CommonFactoryContext. Comparing to ListenerFactoryContext, ServerFactoryContext has a wider lifetime.
This is the prerequisite to share rds configuration among listeners.

Also experimenting move messageValidateVisitor out of CommonFactoryContext. The benefit is that now Server::Instance perfectly implements CommonFactoryContext.

ServerFactoryContext doesn't implement the function. The test shows it is never invoked.
Will put messageValidateVisitor into FactoryContext: this piece of work will be done in follow up PR

This PR is extracted from #8454

Fixes #8303

Signed-off-by: Yuchen Dai <silentdai@gmail.com>