Initialize cluster manager for runtime in validation check (#8986)
In validation mode, the runtime loader's cluster manager was null, leading to a segfault when validation was run on configs with rtds_layer specified. This change initializes the runtime loader with the same cluster manager used in the rest of config validation.

Risk Level: Low
Testing: I tested this by building envoy locally and running the binary on a config containing rtds_layer. Without the line I added in source/server/config_validation/server.cc, running Envoy in validation mode on the config leads to a segfault. With it, it returns 'OK'.
I updated an existing test file as well to show that the rtds_layer gets initialized and parsed correctly; however, the initialize function in server_test.cc behaves differently from the initialize function in server.cc and will succeed without my change.
Docs Changes: N/A
Release Notes: included

Fixes #8977

Signed-off-by: Lisa Lu <lisalu@lyft.com>