test: fix flakey listener_impl_test (#9132)
The DisableAndEnableListener test creates a real listener and client
connection to test listener enable/disable functionality. Prior to
was not accepted by a disabled listener within 2 seconds. After #7165,
the timer and sleep execute against an `Event::SimulatedTimeSystem`
(e.g. the sleep call increments simulated time and the timer fires
almost immediately). The timer event closes the client connection. If
that occurs before the listener has a pending connection, the
subsequent `listener.enable()` call does not trigger
`listener.onAccept` and the test times out. This seems to happen
primarily on macOS, though in principal the race exists on Linux as
well. However, it may be that in practice localhost connections in
Linux are rarely or never deferred.

Risk Level: low
Testing: 1000x iterations
Doc Changes: n/a
Release Notes: n/a

Signed-off-by: Stephan Zuercher <zuercher@gmail.com>