api: link to previous message type package in API BUILD files. (#9210)
* api: link to previous message type package in API BUILD files.

We need to include the descriptors from the previous message version in
the build. We opt to do this transitively; when you include v3 of a
package, you get the v2 via a transitive dep. This should work based on
alwayslink semantics for cc_library.

The computation of the deps is based on the previous_message_type
annotation, which will allow cross package migrations.

Part of #8082.

Risk level: Low
Testing: Disabled ip_tagging v2 descriptor hack, observed
  version_integration_test. After the BUILD changes, this now passes
  again.

Signed-off-by: Harvey Tuch <htuch@google.com>