ip tagging: remove TODO comment that won't be addressed (#9273)
So clearing the route cache when there's an ip tags hit is something
that we've been doing for ~2 years. We used it in prod, it works and
it's not that expensive.

So, lets get rid of the comment that suggests it's either too
expensive or that it'll be refactored. I don't think the latter --
changing how the route cache works -- will happen in the near future
so lets just remove to the comment to avoid confusing newcomers
or casual readers.

Signed-off-by: Raul Gutierrez Segales <rgs@pinterest.com>