Make only once abort or delay decision (#9597)
In the original logic of fault filter, when the header contains the x-envoy-downstream-service-cluster, fault filter will execute the decision whether to abort or delay twice. As a result, the actual abort or delay ratio does not meet expectations. The expectations is p and the actual value will be 2p-p^2.

For example, when the abort ratio is configured as 50%, if the header contains x-envoy-downstream-service-cluster, the actual abort ratio is 75%.

This problem can be reproduced by enabling add_user_agent. The current PR fixes this problem.

Risk Level: Low
Testing: unit tests updated
Docs Changes: n/a
Release Notes: added

Signed-off-by: wbpcode <comems@msn.com>