test flake: fix flake in ApiListener integration test (#9808)
Description: previously the test was not waiting for the expectation on the server's thread to complete. Therefore, there was a use after free race condition with the ApiListener's TlsCachingDateProvider. This PR makes it so that the test waits for the expectation to be fulfilled  and thus prevents the race.
Risk Level: low
Testing: ran integration test a few thousand times locally on a linux machine.

Fixes #9746

Signed-off-by: Jose Nino <jnino@lyft.com>