config: remove ApiTypeOracle assert (#9973)
This assert currently validates that the return value is non-null. However, call sites actually [guard against `nullptr`](https://github.com/envoyproxy/envoy/pull/9618/files#diff-17967d56376bf9821edb17b67b8a7e63R228-R230) and treat it as a valid codepath.

In Envoy Mobile, we're hitting this assert because we use the v3 configuration and don't include all the v2 protos when building the library. Removing the v2 fallback protos in this scenario can be a valid optimization.

Given the above example and the fact that callers explicitly guard against `nullptr`, this assert should be removed and this codepath considered valid.

Signed-off-by: Michael Rebello <me@michaelrebello.com>