retries: enable retry budget for concurrent retries (#774)
Description: previously the cluster configuration for circuit breakers was not specified. This meant that Envoy Mobile was using the default of 3 concurrent retries. This is not appropriate, specially because all* active traffic goes through the same cluster. This PR changes the configuration and explicitly sets the retry_budget configuration.
Risk Level: low (more retries will actually be allowed)
Testing: local

Signed-off-by: Jose Nino <jnino@lyft.com>
Signed-off-by: JP Simard <jp@jpsim.com>