dns_filter: Add external resolution (#11384)
Extend the DNS Filter so that it can resolve queries using external resolvers
Additional Description: The DNS Filter up to this point resolves queries from its local configuration. This change adds the ability to define external name servers and refer queries that cannot be answered from the local config. We also add metrics to gauge the filters behavior and performance.

Signed-off-by: Alvin Baptiste <alvinsb@gmail.com>