Subject: bugfix potential memory bug: use after free (#11701)
This line of code is returning a local copy of string and try to construct a string_view object with that local copy. As string_view is non-owning, it will potentially cause a memory use after free issue. This is found by clang when in C++17 absl::string_view uses std::string_view.

Signed-off-by: Yifan Yang <needyyang@google.com>