CacheFilter: prepares to store and serve varied responses. (#12763)
Commit Message: CacheFilter: prepares to store and serve varied responses

Additional Description:
Enables the cache filter to store and serve responses containing an accepted Vary header. As of now, a temporary placeholder header is the only accepted header name. In a following PR, we will allow the user to populate an allowlist to configure which headers should be accepted in Vary.

Risk Level: Low
Testing: Unit testing
Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Caio <caiomelo@google.com>