[Windows] Improve fileReadToEnd on Windows (#14447)
See #14413

The problem is demonstrated in `eds_integration_test` but it is actually a production code issue for Windows. The scenario is the following:

Thread 1: `MoveFileEx` from E:\temp/eds.update.pb_text -> E:\temp/eds.pb_text
Thread 2: `ReadDirectoryChangeW` notification triggers and notifies `filesystem_subscription_impl`
Thread 3: `filesystem_subscription_impl` tries to read file `E:\temp/eds.pb_text` and fails
Thread 1: `MoveFileEx` completes

Additional Description:

The way MoveFile is implemented is by opening a handle to the source, issuing a rename command, then closing the handle. 

FILE_SHARE_DELETE is named very badly, but it indicates that the opener is willing to allow the file’s name to change while the handle is open.  It’s named very badly because this can’t happen due to delete, because delete in NT occurs on last handle close, so if you have a handle to a file, the file is going to remain there.  In practice, this flag controls two things: first, it allows the file to be renamed; second, it allows another thread to indicate that it wants to delete the file when the last handle is closed. 

I have tested the following tests before removing the flakes:

1. `tap_filter_integration_test` tested 12k times with no flakes
2. `eds_integration_test` tested 2k times
3. `load_stats_integration_test` tested 300 times

Signed-off-by: Sotiris Nanopoulos <sonanopo@microsoft.com>