config: fix crash when xDS endpoints receive unexpected typed config (#14709)
When a listener config contains an Any proto with an unknown proto type, the config
is rejected. If debug logging is turned on, though, Envoy first attempts to print the
JSON representation of the config. This caused a RELEASE_ASSERT to trigger,
crashing Envoy. This PR renames the unsafe proto-to-JSON conversion function
and replaces its usage with the safe version in the listener manager to prevent
the crash.

Signed-off-by: Alex Konradi <akonradi@google.com>