build: fix libtool label for bazel 3.7+ (#1234)
Description: This tool was meant to be a bazel implementation detail so its public
accessor was removed in https://github.com/bazelbuild/bazel/commit/f8b51ffc2a8a953f64657ae4a21c9e1f1793dae2

Using the cc API from bazel we can dynamically get this. This is still a
bit of a hack because we're writing our own args for this. We might be
able to fetch the args, based on bazel's crosstool, dynamically as well,
but ideally we remove this rule instead.

Fixes https://github.com/envoyproxy/envoy-mobile/issues/1233

Signed-off-by: Keith Smiley <keithbsmiley@gmail.com>
Signed-off-by: JP Simard <jp@jpsim.com>