stats: add apis for custom tags (#1266)
Description:

This PR adds APIs for custom tags for both iOS and Android, and for all four types of stats.

Two APIs are supported for each type of stats: 1) stats creation with tags; 2) creating a new stats from an existing stats with a list of tags attached.

The first API is designed for stats creation during compile time, calling the first API to create stats during runtime is not restricted. The second API is designed for stats creation during runtime, similar to the first API, calling this API to create stats during compile time is not restricted.

`Tag` classes do not enforce checks like `Element` classes, the idea is to leave the implementations calling this set of APIs to enforce their own system/platform specific restrictions on tag keys and tag values.

Risk Level: Medium

Testing: unit tests, deployed to Lyft iOS and Android apps for smoke tests

Signed-off-by: Jingwei Hao jingwei.hao@gmail.com
Signed-off-by: JP Simard <jp@jpsim.com>