http3: improving upstream TLS code (#16281)
Addressing a TODO for upstream quic crypto. The clientContextConfig looks safe to access, as it does not reload, but QUIC was using the context to create a fresh Tls::ClientContextImpl per connection instead of reusing the factory built shared Ssl::ClientContextSharedPtr. This PR moves it to the ssl shared pointer.

Risk Level: low (mostly H3 only)
Testing: n/a (refactor)
Docs Changes: n/a
Release Notes: n/a
part of #15649

Signed-off-by: Alyssa Wilk <alyssar@chromium.org>