Add "Flow" Integration Test for Android (#1446)
This test helps documenting what is the Envoy-Mobile flow logic. This is useful in regards with Cronvoy development, however, this is not testing anything related to Cronvoy - it is a pure Envoy-Mobile test.

Description: Add tests exercising some basic HTTP flows with the Android library
Risk Level: Inexistent
Testing: This PR only adds a new test - no prod code is being changed
Docs Changes: N/A
Release Notes: N/A
Signed-off-by: Charles Le Borgne <cleborgne@google.com>
Signed-off-by: JP Simard <jp@jpsim.com>