engine: ensure thread safety in postinit setup (#1471)
Description: It's critical that postinit setup is thread-safe relative to the dispatcher, but we can't actually call Event::Dispatcher::isThreadSafe() because it blows up at this juncture. This safety check is leveraged elsewhere by Envoy to execution is happening on the main thread regardless of dispatch.
Risk Level: Low
Testing: Local and CI

Signed-off-by: Mike Schore <mike.schore@gmail.com>
Signed-off-by: JP Simard <jp@jpsim.com>